refactor(import): use import insted of require
