+++ packages/s2s-helper-get-initial-state/src/index.test.js
@@ -1,5 +1,5 @@
 // @flow
-const helper = require('./').default
+import helper from '.'
 
 test('get initalState', () => {
   const fixture = `
